backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-core==0.1.0
langchain-text-splitters==0.0.1
tavily-python==0.3.0
faiss-cpu==1.9.0.post1
pymupdf>=1.24.0
python-dotenv==1.0.0
aiofiles==23.2.1

